Refactor GET provider by id
Jira: MODKBEKBJ-109
Changed getEholdingsProvidersByProviderId to use RMAPITemplate
